move group dropdown
